Delete FIELDBUS_WORD_TO_PERCENT.fbt
This commit deletes the file FIELDBUS_WORD_TO_PERCENT.fbt, which was used to convert a hexadecimal percent to a REAL value. The file contained identification information, version info, compiler info, interface list, and an algorithm for converting the input values. The algorithm checked if the input qualifier was TRUE and if the input value was less than or equal to a predefined maximum value. If both conditions were met, it assigned the input value to an output variable and calculated the output in percent as a REAL value.