Migrates to logiBUS I/O modules
Updates the application to use logiBUS I/O modules for digital input and output.

This change replaces the generic IX and QX FB types with the more specific logiBUS_IX and logiBUS_QX types.
It also introduces parameters to map the input and output to specific logiBUS hardware.
The old mapping files and obsolete SUB files are removed.