Configures traffic light sequence with logiBUS
Adapts the traffic light control sequence to use logiBUS components for input and output.

This change replaces generic QX and IE function blocks with logiBUS specific implementations (logiBUS_QX and logiBUS_IE) for better integration with the target hardware.
It configures the output and input parameters to specific logiBUS channels.

Additionally, it adjusts the positions of some function blocks and subapps for better layout.