Configures sub-application for data panel control
Configures a sub-application to map Funk keys to DataPanel outputs.

This commit sets up a sub-application (`Uebung_003b2`) which connects digital input from a Funk device to digital outputs on a DataPanel, using an intermediary FB `DataPanel_MI_QX`. The configuration includes specifying the input key, SA member, and output for each of the 15 sub-applications.