Updates CAN message handling for isobus
Refactors CAN message data type to include size and priority,
and adds timestamp to received messages. This improves the
handling of CAN bus communication in the isobus system by
providing additional information and context for each message.
The default priority for PGNs is set to 7. Changes some
references of filter names in a SUB file.