Adds package name to compiler info
Adds the package name "Uebungen" to the compiler information of all SUB application files.

This change ensures that the 4diac IDE correctly identifies and compiles the application files within the "Uebungen" package.

Updates sub-application types to include the package name.

This ensures that the 4diac IDE correctly identifies and compiles the application files within the "Uebungen" package.